validate required items definition on command handler

+++ nts.uk/uk.com/com.ctx/bs/person/nts.uk.ctx.bs.person.app/src/main/java/command/layout/NewLayoutCommandHandler.java
@@ -7,6 +7,9 @@ import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.transaction.Transactional;
 
+import find.person.info.item.PerInfoItemDefFinder;
+import nts.arc.error.BusinessException;
+import nts.arc.error.RawErrorMessage;
 import nts.arc.layer.app.command.CommandHandler;
 import nts.arc.layer.app.command.CommandHandlerContext;
 import nts.uk.ctx.bs.person.dom.person.layout.INewLayoutReposotory;
@@ -29,12 +32,25 @@ public class NewLayoutCommandHandler extends CommandHandler<NewLayoutCommand> {
 	@Inject
 	ILayoutPersonInfoClsDefRepository clsDefRepo;
 
+	@Inject
+	PerInfoItemDefFinder itemDefFinder;
+
 	@Override
 	protected void handle(CommandHandlerContext<NewLayoutCommand> context) {
 
 		// get new layout domain and command
 		NewLayout update = layoutRepo.getLayout().get();
 		NewLayoutCommand command = context.getCommand();
+
+		// validate all usecase [Registration] at here
+		// throw exception if not valid
+		List<String> requiredIds = itemDefFinder.getRequiredIds();
+		List<String> allSaveItemIds = command.getItemsClassification().stream().map(m -> m.listItemClsDf)
+				.flatMap(List::stream).map(m -> m.getPersonInfoItemDefinitionID()).collect(Collectors.toList());
+
+		if (!allSaveItemIds.containsAll(requiredIds)) {
+			throw new BusinessException(new RawErrorMessage("Msg_201"));
+		}
 
 		// rmove all classification in this layout
 		classfRepo.removeAllByLayoutId(update.getLayoutID());
